requirements.txt
-e svn+http://code.djangoproject.com/svn/django/trunk/#egg=django-trunk
pillow==1.7.6
South==0.7.3
django-debug-toolbar==0.9.4
django-filebrowser==3.4.0
django-grappelli==2.3.5
django-mptt==0.5.2
django-sitetree==0.6
django-tinymce==1.5.1b2
wsgiref==0.1.2
django-extensions
django-celery==2.4.2
django-kombu==0.9.4
celery
